+++ src/cobalt/kissmark.py
@@ -1,0 +1,36 @@
+# -*- coding: utf-8 -*-
+"""Story: The kissmark
+"""
+import os
+import sys
+sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
+sys.path.append('storybuilder')
+
+from storybuilder.builder import world as wd
+from src.cobalt import config as cnf
+THM = cnf.THEMES
+
+# outline
+def story_baseinfo(w: wd.World):
+    return [
+            ]
+
+def story_outline(w: wd.World):
+    return [
+            ]
+
+# main
+def story(w: wd.World):
+    return (w.maintitle("キスマーク"),
+            )
+
+def main(): # pragma: no cover
+    from src.cobalt.story import world
+    w = world()
+    return w.build(story(w))
+
+
+if __name__ == '__main__':
+    import sys
+    sys.exit(main())
+