+++ src/cobalt/zebra.py
@@ -7,7 +7,7 @@ sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
 sys.path.append('storybuilder')
 
 from storybuilder.builder import world as wd
-from src.zebra import config as cnf
+from src.cobalt import config as cnf
 THM = cnf.THEMES
 
 
@@ -557,12 +557,6 @@ def story_outline(w: wd.World):
             ]
 
 # main
-def world():
-    w = wd.World("The zebra-bra project")
-    w.set_db(cnf.CHARAS, cnf.STAGES, cnf.DAYS, cnf.ITEMS, cnf.INFOS, cnf.FLAGS,
-            )
-    return w
-
 def story(w: wd.World):
     return (w.maintitle("ぜぶらぶら"),
             ep_intro(w),
@@ -576,6 +570,7 @@ def story(w: wd.World):
             )
 
 def main(): # pragma: no cover
+    from src.cobalt.story import world
     w = world()
     return w.build(story(w))