+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Re-Project: Kissmark
 ### Changed
 
 ## [0.6.0] - 2019-10-01